+++ sample-app/src/components/CometChatDetails/CometChatThreadedMessages.tsx
@@ -56,29 +56,42 @@ export const CometChatThreadedMessages = (props: ThreadProps) => {
                             user={(selectedItem as CometChat.Conversation)?.getConversationType?.() === "user" ? (selectedItem as CometChat.Conversation)?.getConversationWith() as CometChat.User : (selectedItem as CometChat.User).getUid?.() ? selectedItem as CometChat.User : undefined}
                             group={(selectedItem as CometChat.Conversation)?.getConversationType?.() === "group" ? (selectedItem as CometChat.Conversation)?.getConversationWith() as CometChat.Group : (selectedItem as CometChat.Group).getGuid?.() ? selectedItem as CometChat.Group : undefined}
                         />
-                    </div> : <div className="message-composer-blocked" onClick={()=>{
-                        let user: CometChat.User | null = null;
+                    </div> : 
+                    (
+                    <div className="message-composer-blocked">
+                        <div className="message-composer-blocked__text">
+                            {getLocalizedString("cannot_send_to_blocked_user")}{" "}
+                            <a
+                                onClick={() => {
+                                let user: CometChat.User | null = null;
 
-                        if (selectedItem instanceof CometChat.User) {
-                            user = selectedItem;
-                        } else if (
-                            selectedItem instanceof CometChat.Conversation &&
-                            selectedItem.getConversationType() === CometChat.RECEIVER_TYPE.USER &&
-                            selectedItem.getConversationWith() instanceof CometChat.User
-                        ) {
-                            user = selectedItem.getConversationWith() as CometChat.User;
-                        }
-                        if (user) {
-                            CometChat.unblockUsers([user.getUid()]).then(() => {
-                                user?.setBlockedByMe(false);
-                                CometChatUserEvents.ccUserUnblocked.next(user as CometChat.User);
-                            });
-                        }
-            }}>
-               <div className="message-composer-blocked__text">
-                {getLocalizedString("cannot_send_to_blocked_user")} <a>   {getLocalizedString("click_to_unblock")}</a>
-               </div>
-            </div>}
+                                if (selectedItem instanceof CometChat.User) {
+                                    user = selectedItem;
+                                } else if (
+                                    selectedItem instanceof CometChat.Conversation &&
+                                    selectedItem.getConversationType() ===
+                                    CometChat.RECEIVER_TYPE.USER &&
+                                    selectedItem.getConversationWith() instanceof
+                                    CometChat.User
+                                ) {
+                                    user =
+                                    selectedItem.getConversationWith() as CometChat.User;
+                                }
+                                if (user) {
+                                    CometChat.unblockUsers([user.getUid()]).then(() => {
+                                    user?.setBlockedByMe(false);
+                                    CometChatUserEvents.ccUserUnblocked.next(
+                                        user as CometChat.User
+                                    );
+                                    });
+                                }
+                                }}
+                            >
+                                {getLocalizedString("click_to_unblock")}
+                            </a>
+                        </div>
+                    </div>
+                    )}
                 </>}
         </div>
     );